made the addmovie page a folder and started separating into smaller chunks
